fix wrong filename and one file in wrong package
